backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.12.1
boto3==1.29.3
pytest==7.4.3
httpx==0.25.2
python-dotenv==1.0.0
mangum==0.17.0
aws-cdk-lib==2.110.0
constructs>=10.0.0
aws-cdk.aws-lambda-python-alpha==2.110.0a0
email-validator==2.1.0 